Fix remaining cross-imported utils.
